Finalização da home page, implementação de controller para aba de pedidos.
