--- src/main/java/br/com/danichs/projetomvc/controller/HomeController.java
@@ -1,28 +1,0 @@
-package br.com.danichs.projetomvc.controller;
-
-import java.util.Arrays;
-import java.util.List;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-
-import br.com.danichs.projetomvc.model.Pedido;
-
-@Controller
-public class HomeController {
-
-    @GetMapping("/home")
-    public String home(Model model) {
-        Pedido pedido = new Pedido();
-        pedido.setNomeProduto("Xiaomi Redmi Note 8");
-        pedido.setUrlImagem("https://m.media-amazon.com/images/I/51RIqTiRc8L.__AC_SX300_SY300_QL70_ML2_.jpg");
-        pedido.setUrlProduto("https://www.amazon.com.br/Smartphone-Xiaomi-Redmi-Note-Camera/dp/B07Z5BBG56");
-        pedido.setDescricao("uma descrição qualquer para esse pedido");
-
-        List<Pedido> pedidos = Arrays.asList(pedido);
-        model.addAttribute("pedidos", pedidos);
-
-        return "home";
-    }
-}
