+++ src/main/java/br/com/danichs/projetomvc/springmvc/SpringmvcApplication.java
@@ -2,10 +2,8 @@ package br.com.danichs.projetomvc.springmvc;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
-@ComponentScan(basePackages = {"br.com.danichs.projetomvc.controller"})
 public class SpringmvcApplication {
 
 	public static void main(String[] args) {
