+++ src/main/java/br/com/danichs/projetomvc/springmvc/controller/HomeController.java
@@ -1,0 +1,27 @@
+package br.com.danichs.projetomvc.springmvc.controller;
+
+import java.util.List;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+
+import br.com.danichs.projetomvc.springmvc.model.Pedido;
+import br.com.danichs.projetomvc.springmvc.repository.PedidoRepository;
+
+@Controller
+public class HomeController {
+
+    @Autowired
+    private PedidoRepository pedidoRepository;
+
+    @GetMapping("/home")
+    public String home(Model model) {
+
+        List<Pedido> pedidos = pedidoRepository.findAll();
+        model.addAttribute("pedidos", pedidos);
+
+        return "home";
+    }
+}
