+++ src/main/java/br/com/danichs/projetomvc/springmvc/repository/PedidoRepository.java
@@ -1,0 +1,11 @@
+package br.com.danichs.projetomvc.springmvc.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import br.com.danichs.projetomvc.springmvc.model.Pedido;
+
+@Repository
+public interface PedidoRepository extends JpaRepository<Pedido, Long> {
+
+}
