+++ src/main/resources/application.properties
@@ -2,6 +2,6 @@ spring.datasource.url=jdbc:mysql://localhost:3306/springmvcdb
 spring.datasource.username=root
 spring.datasource.password=root
 
-spring.jpa.show-sql: true
+spring.jpa.show-sql = true
 spring.jpa.hibernate.ddl-auto=update
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
