+++ src/main/resources/templates/home.html
@@ -8,6 +8,18 @@
 </head>
 <body>
     <div class="container">
+        <div class="mt-4 p-5 bg-body-secondary mb-2 rounded">
+            <h1 class="display-4">Meus Pedidos</h1>
+        </div>
+        <nav class="navbar navbar-expand navbar-light bg-light d-flex justify-content-between mb-3 p-2 rounded">
+            <div class="navbar-nav">
+                <a class="nav-item nav-link active" href="">Todos</a>
+                <a class="nav-item nav-link" href="">Aguardando</a>
+                <a class="nav-item nav-link" href="">Aprovado</a>
+                <a class="nav-item nav-link" href="">Entregue</a>
+            </div>
+            <a href="/pedido/formulario"><button class="btn btn-outline-secondary my-2 my-sm-0">Novo</button></a>
+        </nav>
         <div class="card" th:each="pedido : ${pedidos}">
             <div class="card-header" th:text="${pedido.nomeProduto}">Nome do produto</div>
             <div class="card-body">
